src/stac_ingestion/stac_collection/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
azure-common==1.1.28
azure-core==1.25.0
azure-identity==1.10.0
azure-servicebus==7.8.0
azure-storage-blob==12.13.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
frozenlist==1.3.1
google-api-core==2.8.2
google-auth==2.11.0
googleapis-common-protos==1.56.4
idna==3.3
isodate==0.6.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.2
oauthlib==3.2.0
opencensus==0.11.0
opencensus-context==0.1.3
opencensus-ext-azure==1.1.7
portalocker==2.5.1
protobuf==4.21.5
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.3.0
uamqp==1.6.0
urllib3==1.26.12
yarl==1.8.1
